Add files via upload
Fixed and random effects for m2 with only 2000 and 2001 in the dataset (i1 = 2), just to get an idea of how close the estimates are so to test packages agreement. Despite the evident agreement, it looks like INLA tends to deviate from others (due to being an approximation?), which is especially notable for gen_env.